Move LSP scope tests outside src
